add multi-threaded output stream parsing, presets, removed RNPxl
